+++ tests/test_checker/test_noqa.py
@@ -163,6 +163,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS354': 1,
     'WPS355': 1,
     'WPS356': 1,
+    'WPS357': 0,  # logically unacceptable.
 
     'WPS400': 0,  # defined in ignored violations.
     'WPS401': 0,  # logically unacceptable.