+++ CHANGELOG.md
@@ -29,7 +29,8 @@ Semantic versioning in our case means:
 - Adds domain names options:
   `--allowed-domain-names` and `--forbidden-domain-names`,
   that are used to create variable names' blacklist #1106
-- Forbids to use `:=` operator, it reuses `WPS332` code
+- Forbids to use `\r` (carriage return) as line breaks in strings #1111
+- Forbids to use `:=` operator, it now reuses `WPS332` code
 - Forbids to use positional only `/` arguments
 - Forbids to have too many names imported from a single `from ... import`
 - Adds `__call__` to list of methods that should be on top #1125